apps-write-back/pyproject.toml
[project]
name = "apps-data-entry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=3.0.4",
    "dash-ag-grid>=31.3.1",
    "dash-iconify>=0.1.2",
    "dash-mantine-components>=2.0.0",
    "databricks-sdk>=0.56.0",
    "databricks-sql-connector>=4.0.3",
    "psycopg[binary]>=3.2.9",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
